fix(new-design): Separate app.js
Separate alpine components from app.js to separeted files
