+++ resources/js/alpine/crudTable.js
@@ -1,0 +1,33 @@
+export default () => ({
+
+  actionsOpen: false,
+  actions (type) {
+    let all = this.$root.querySelector('.actionsAllChecked')
+
+    if (all === null) {
+      return
+    }
+
+    let checkboxes = this.$root.querySelectorAll('.tableActionRow')
+    let ids = document.querySelectorAll('.actionsCheckedIds')
+
+    let values = []
+
+    for (let i = 0, n = checkboxes.length; i < n; i++) {
+      if (type === 'all') {
+        checkboxes[i].checked = all.checked
+      }
+
+      if (checkboxes[i].checked && checkboxes[i].value) {
+        values.push(checkboxes[i].value)
+      }
+    }
+
+    for (let i = 0, n = ids.length; i < n; i++) {
+      ids[i].value = values.join(';')
+    }
+
+    this.actionsOpen = !!(all.checked || values.length)
+  },
+
+})
