+++ resources/js/alpine/pivot.js
@@ -1,0 +1,16 @@
+/* Pivot */
+
+export default () => ({
+
+  autoCheck () {
+    let checker = this.$root.querySelector('.pivotChecker')
+    let fields = this.$root.querySelectorAll('.pivotFields')
+
+    fields.forEach(function (value, key) {
+      value.addEventListener('input', (event) => {
+        checker.checked = event.target.value
+      })
+    })
+  },
+
+})
