+++ resources/js/alpine/tooltip.js
@@ -1,0 +1,12 @@
+/* Tooltip */
+
+import tippy from 'tippy.js'
+
+export default (text, config = {}) => ({
+
+  tooltipInstance: null,
+  init() {
+    this.tooltipInstance = tippy(this.$el, { ...config, content: text })
+  },
+
+})
